# Conflicts:
#	src/main/java/com/fitj/controllers/clients/ControllerClientDetail.java
#	src/main/java/com/fitj/facades/FacadeClient.java
#	src/main/java/com/fitj/facades/FacadeCoach.java
#	src/main/java/com/fitj/facades/FacadePack.java
#	src/main/java/com/fitj/facades/FacadeProgrammeNutrition.java
#	src/main/java/com/fitj/facades/FacadeProgrammeSportif.java
#	src/main/java/com/fitj/facades/FacadeSeance.java

+++ src/main/java/com/fitj/dao/postgresql/DAOPaiementPostgreSQL.java
@@ -58,7 +58,7 @@ public class DAOPaiementPostgreSQL extends DAOPaiement {
                 Map<String, Object> data = listData.get(i);
                 paiements.add(new Paiement(
                         ((Long)data.get("id")).intValue(),
-                        ((Long)data.get("montant")).intValue(),
+                        (double)data.get("montant"),
                         PaiementType.getPaiementType((String)data.get("type"))));
                 i++;
             }
